Add a fake terminal chat room
This commit adds a version of RoomProxy that fakes a chat room using
the terminal. The fake chat room will accept message from stdin and
send responses to stdout, but will act as if no users are in the room.
There are also some modifications in other files to make the API between
the SE chat RoomProxy and the terminal RoomProxy compatible enough for
the script to work. run.py now uses the terminal room by default, or
if the room ID in the config file is explicitly set as 0 or 'terminal'.