Allow the terminal chat to pretend that mods are present or pingable
This commit gives the terminal chat emulator the ability to pretend that
certain moderators are present in the chat room, or not present but pingable.
The sets of present and pingable mods are given to the constructor; obviously
it doesn't reflect their actual status on the SE chat system. The driver
program reads options from the config file to initialize these values.