Update bot to account for unknown mod IDs
This commit updates the bot code to account for moderators whose chat
profile ID is unknown (and thus is recorded in the JSON file as -1).
In most cases these moderators will be ignored for the purpose of
pinging, but in a whois command, the bot will append a list of
moderators whose chat profile IDs are unknown.