Create a class to hold moderator info
This commit creates a new class to hold the moderator information,
and changes the code to create an instance of that class and pass it
around. With this change, there is no longer a global variable that
stores the information. Changing from a global variable to a custom
class makes it easier to refactor the code in the future, perhaps
to track different groups of chat users or other changes.